chore: utilize Router in App.js, move the map view content to /src/pages/HomePage.js)
